+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Sync client-side schema with server-side schema ([#10](https://github.com/provablyfine/pf/issues/10))
 - Align key thumbprint calculation with RFC 6738 ([#16](https://github.com/provablyfine/pf/issues/16))
 - Audit session duration ([#8](https://github.com/provablyfine/pf/issues/8))
+- Allow login when we do not have a working ssh-agent ([#6](https://github.com/provablyfine/pf/issues/6))
 
 ## 0.3.0 - 2026-06-05